# docs/openapi.yaml
openapi: 3.0.1
info:
  description: "REST API for managing gym class sessions, memberships, bookings and\
    \ waitlists."
  title: Fitness Class Booking System
  version: "1.0"
servers:
- description: Default server
  url: /
...
paths:
...
  /api/v1/members/{memberId}/memberships/{membershipId}:
    get:
      description: "Retrieves a single membership by its ID. Returns effective status\
        \ (e.g., ACTIVE→EXPIRED when expiry has passed, PENDING→ACTIVE when start\
        \ time has passed). The stored row is not modified."
      operationId: getMembership
      parameters:
      - in: path
        name: memberId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: membershipId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipResponse"
          description: Membership retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Membership not found
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Invalid membership ID format
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Retrieve a membership by ID
      tags:
      - Memberships
...
  /api/v1/memberships/{membershipId}/ledger:
    get:
      description: "Returns the append-only credit transaction ledger for a membership,\
        \ including the current balance reconciliation. For unlimited memberships,\
        \ returns empty content with null credit figures. Transactions are returned\
        \ in oldest-first order (by creation time)."
      operationId: getLedger
      parameters:
      - in: path
        name: membershipId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLedgerResponse"
          description: Ledger retrieved successfully
        "404":
          content:
            application/json: {}
          description: Membership not found
        "422":
          content:
            application/json: {}
          description: Invalid membership ID format
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Retrieve credit transaction ledger for a membership
      tags:
      - Credit Ledger
...
components:
  schemas:
...
    CreditLedgerResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        credits_initial:
          type: integer
          format: int32
        credits_remaining:
          type: integer
          format: int32
        reconciles:
          type: boolean
    Entry:
      type: object
      properties:
        balance_after:
          type: integer
          format: int32
        created_at:
          type: string
        delta:
          type: integer
          format: int32
        id:
          type: string
        reason:
          type: string
    ErrorEnvelope:
      type: object
      properties:
        code:
          type: string
          enum:
          - MALFORMED_REQUEST
          - NOT_PERMITTED
          - MEMBER_SUSPENDED
          - MEMBER_INACTIVE
          - NOT_FOUND
          - MEMBER_NOT_FOUND
          - MEMBERSHIP_PLAN_NOT_FOUND
          - MEMBERSHIP_NOT_FOUND
          - CLASS_TYPE_NOT_FOUND
          - INSTRUCTOR_NOT_FOUND
          - ROOM_NOT_FOUND
          - SESSION_NOT_FOUND
          - BOOKING_NOT_FOUND
          - WAITLIST_ENTRY_NOT_FOUND
          - CONFLICT
          - CONCURRENT_MODIFICATION
          - DUPLICATE_BOOKING
          - SESSION_FULL
          - SESSION_CANCELLED
          - SESSION_ALREADY_CANCELLED
          - SESSION_NOT_BOOKABLE
          - BOOKING_ALREADY_CANCELLED
          - BOOKING_NOT_CANCELLABLE
          - BOOKING_ALREADY_CHECKED_IN
          - CHECK_IN_WINDOW_NOT_OPEN
          - WAITLIST_ALREADY_JOINED
          - WAITLIST_ALREADY_PROCESSED
          - MEMBERSHIP_ALREADY_ACTIVE
          - CREDITS_INSUFFICIENT
          - MEMBERSHIP_NO_CREDITS
          - OVERLAPPING_BOOKING
          - BOOKING_OVERLAPS_EXISTING
          - LATE_CANCEL_NO_REFUND
          - INSTRUCTOR_SCHEDULE_CONFLICT
          - ROOM_SCHEDULE_CONFLICT
          - CLASS_TYPE_NAME_ALREADY_EXISTS
          - CLASS_TYPE_INACTIVE
          - INSTRUCTOR_EMAIL_ALREADY_EXISTS
          - INSTRUCTOR_INACTIVE
          - INSTRUCTOR_HAS_FUTURE_SESSIONS
          - ROOM_NAME_ALREADY_EXISTS
          - ROOM_INACTIVE
          - ROOM_HAS_FUTURE_SESSIONS
          - SESSION_INSTRUCTOR_CONFLICT
          - SESSION_ROOM_CONFLICT
          - SESSION_CAPACITY_EXCEEDS_ROOM
          - IDEMPOTENCY_KEY_CONFLICT
          - JOB_ALREADY_RUNNING
          - MEMBER_EMAIL_ALREADY_EXISTS
          - MEMBER_ALREADY_SUSPENDED
          - MEMBER_NOT_SUSPENDED
          - PLAN_NAME_ALREADY_EXISTS
          - PLAN_ALREADY_INACTIVE
          - PLAN_INACTIVE
          - MEMBERSHIP_ALREADY_QUEUED
          - MEMBERSHIP_START_BEFORE_CURRENT_EXPIRY
          - MEMBERSHIP_NOT_CANCELLABLE
          - SESSION_NOT_EDITABLE
          - SESSION_ALREADY_STARTED
          - SESSION_CAPACITY_BELOW_BOOKED
          - VALIDATION_FAILED
          - OUT_OF_RANGE
          - INVALID_DATE_RANGE
          - INVALID_PAGINATION
          - INVALID_SORT_FIELD
          - INVALID_FORMAT
          - INVALID_ENUM
          - UNKNOWN_FIELD
          - FUTURE_REQUIRED
          - INVALID_RANGE
          - RECURRENCE_EMPTY
          - INTERNAL_ERROR
        detail:
          type: string
        errors:
          type: array
          description: Per-field validation errors. Present ONLY on 422 responses.
          items:
            $ref: "#/components/schemas/FieldError"
        instance:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        title:
          type: string
        traceId:
          type: string
        type:
          type: string
    FieldError:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: object
...
    MembershipResponse:
      type: object
      description: Membership response
      properties:
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
          example: 2026-09-23T10:15:30.123Z
        creditsInitial:
          type: integer
          format: int32
          description: Initial credits at assignment (null for unlimited memberships)
          example: 10
        creditsRemaining:
          type: integer
          format: int32
          description: Remaining credits (null for unlimited memberships)
          example: 10
        expiresAt:
          type: string
          format: date-time
          description: Membership expiry time (UTC)
          example: 2026-10-23T10:00:00Z
        id:
          type: string
          format: uuid
          description: Membership unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        memberId:
          type: string
          format: uuid
          description: Member unique identifier
          example: 550e8400-e29b-41d4-a716-446655440001
        planId:
          type: string
          format: uuid
          description: Membership plan unique identifier (snapshotted at assignment
            time)
          example: 550e8400-e29b-41d4-a716-446655440002
        startsAt:
          type: string
          format: date-time
          description: Membership start time (UTC)
          example: 2026-09-23T10:00:00Z
        status:
          type: string
          description: "Membership status (ACTIVE, PENDING, EXPIRED, CANCELLED)"
          example: ACTIVE
        unlimited:
          type: boolean
          description: True if this is an unlimited-credits membership (creditsRemaining
            is null)
          example: false
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: 2026-09-23T10:15:30.123Z
        version:
          type: integer
          format: int64
          description: Optimistic lock version
          example: 0
...